refactor: arranging files, commenting editing controller
