+++ lib/src/view/typeset_editing_controller.dart
@@ -1,23 +1,23 @@
-import 'package:flutter/material.dart';
-import 'package:typeset/src/core/typeset_parser.dart';
+// import 'package:flutter/material.dart';
+// import 'package:typeset/src/core/typeset_parser.dart';
 
-///
-class TypesetEditingController extends TextEditingController {
-  ///
+// ///
+// class TypesetEditingController extends TextEditingController {
+//   ///
 
-  @override
-  TextSpan buildTextSpan({
-    required BuildContext context,
-    TextStyle? style,
-    required bool withComposing,
-  }) {
-    final spans = TypesetParser.parser(
-      inputText: text,
-    );
+//   @override
+//   TextSpan buildTextSpan({
+//     required BuildContext context,
+//     TextStyle? style,
+//     required bool withComposing,
+//   }) {
+//     final spans = TypesetParser.parser(
+//       inputText: text,
+//     );
 
-    return TextSpan(
-      style: style,
-      children: spans,
-    );
-  }
-}
+//     return TextSpan(
+//       style: style,
+//       children: spans,
+//     );
+//   }
+// }
