+++ lib/typeset.dart
@@ -3,7 +3,7 @@
 library typeset;
 
 export 'src/core/typeset_reserved.dart';
-export 'src/typeset.dart';
-export 'src/view/typeset_editing_controller.dart';
-export 'src/typeset_ext.dart';
+export 'src/view/typeset.dart';
 export 'src/view/typeset_context_menus.dart';
+export 'src/view/typeset_editing_controller.dart';
+export 'src/view/typeset_ext.dart';
